requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
aiohttp==3.12.15
asyncio==4.0.0
python-dotenv==1.1.1